fish_backbone/requirements.txt

paddlepaddle
paddlex
flask
opencv-python
numpy
pillow
